Section-2/pyproject.toml
[project]
name = "ai-course-eden-marco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.34.4",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "numpy>=2.3.2",
    "openai>=1.102.0",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "scikit-learn>=1.7.1",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]
